refactor(core): Extract HTTP request, authentication, and legacy adapter from request-helper-functions
