+++ packages/core/src/execution-engine/node-execution-context/index.ts
@@ -14,7 +14,6 @@ export { constructExecutionMetaData } from './utils/construct-execution-metadata
 export { getAdditionalKeys, getNonWorkflowAdditionalKeys } from './utils/get-additional-keys';
 export { normalizeItems } from './utils/normalize-items';
 export { parseIncomingMessage } from './utils/parse-incoming-message';
-export { parseRequestObject } from './utils/request-helper-functions';
 export { returnJsonArray } from './utils/return-json-array';
 export { resolveSourceOverwrite } from './utils/resolve-source-overwrite';
 export * from './utils/binary-helper-functions';
