+++ packages/core/src/execution-engine/node-execution-context/utils/request-helpers/index.ts
@@ -1,16 +1,18 @@
 export {
+	buildTargetUrl,
 	createFormDataObject,
 	digestAuthAxiosConfig,
 	generateContentLengthHeader,
 	getBeforeRedirectFn,
 	getHostFromRequestObject,
+	getUrlFromProxyConfig,
 	isFormDataInstance,
 	isIgnoreStatusErrorConfig,
+	resolveLegacyRequestUrl,
 	searchForHeader,
-	tryParseUrl,
-	buildTargetUrl,
-	getUrlFromProxyConfig,
 	setAxiosAgents,
+	tryParseUrl,
+	validateUrlSsrf,
 } from './axios-utils';
 export {
 	applyPaginationRequestData,
