+++ packages/core/src/execution-engine/node-execution-context/utils/request-helpers/oauth.ts
@@ -15,8 +15,6 @@ import { AuthError, ClientOAuth2 } from '@n8n/client-oauth2';
 import type { AxiosError } from 'axios';
 import { createHmac } from 'crypto';
 import get from 'lodash/get';
-import isEmpty from 'lodash/isEmpty';
-import { NodeOperationError, ApplicationError, jsonParse } from 'n8n-workflow';
 import type {
 	IAllExecuteFunctions,
 	ICredentialDataDecryptedObject,
@@ -28,19 +26,13 @@ import type {
 	IWorkflowExecuteAdditionalData,
 	Logger as WorkflowLogger,
 } from 'n8n-workflow';
+import { ApplicationError, jsonParse, NodeOperationError } from 'n8n-workflow';
 import type { Token } from 'oauth-1.0a';
 import clientOAuth1 from 'oauth-1.0a';
 
 import type { IResponseError } from '@/interfaces';
 
-const NoBodyHttpMethods = ['GET', 'HEAD', 'OPTIONS'];
-
-const removeEmptyBody = (requestOptions: IHttpRequestOptions | IRequestOptions) => {
-	const method = requestOptions.method || 'GET';
-	if (NoBodyHttpMethods.includes(method) && isEmpty(requestOptions.body)) {
-		delete requestOptions.body;
-	}
-};
+import { removeEmptyBody } from './http-request';
 
 function createOAuth2Client(credentials: OAuth2CredentialData): ClientOAuth2 {
 	// Split and trim scopes; empty scope tokens are not RFC 6749-compliant and may be rejected by authorization servers
