+++ packages/nodes-base/nodes/Microsoft/Entra/test/MicrosoftEntra.node.test.ts
@@ -1,6 +1,6 @@
 import { CredentialsHelper } from '@nodes-testing/credentials-helper';
 import { NodeTestHarness } from '@nodes-testing/node-test-harness';
-import { convertN8nRequestToAxios } from 'n8n-core/dist/execution-engine/node-execution-context/utils/request-helper-functions';
+import { convertN8nRequestToAxios } from 'n8n-core/dist/execution-engine/node-execution-context/utils/request-helpers/http-request';
 import type {
 	IExecuteSingleFunctions,
 	ILoadOptionsFunctions,
